Add tests
Add test for CalculateFinalTest function

Add test for MarksAreValid function

- Chain of propmises was optimized in script.js

+++ calculator/calculator.go
@@ -1,6 +1,5 @@
 package calculator
 
-// float finalMark = (float) (currentAverageIntFMLayout * 0.6 + examIntFMLayout * 0.4);
 func CalculateFinal(averageMark, examMark float32) float32 {
-	return (averageMark)*0.6 + float32(examMark)*0.4
+	return (averageMark)*0.6 + (examMark)*0.4
 }
